broker/config.go
package broker

import (
	"errors"
	"fmt"
)

type Config struct {
	Username                     string  `json:"username"`
	Password                     string  `json:"password"`
	TLSCertFile                  string  `json:"tls_cert_file"`
	TLSKeyFile                   string  `json:"tls_key_file"`
	AllowUserProvisionParameters bool    `json:"allow_user_provision_parameters"`
	AllowUserUpdateParameters    bool    `json:"allow_user_update_parameters"`
	AllowUserBindParameters      bool    `json:"allow_user_bind_parameters"`
	Catalog                      Catalog `json:"catalog"`
}

func (c Config) Validate() error {
	if c.Username == "" {
		return errors.New("Must provide a non-empty Username")
	}

	if c.Password == "" {
		return errors.New("Must provide a non-empty Password")
	}

	if err := c.Catalog.Validate(); err != nil {
		return fmt.Errorf("Validating Catalog configuration: %s", err)
	}

	return nil
}
